% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionBayesian.R
\name{Actual.Values}
\alias{Actual.Values}
\title{Actual Values of Variables}
\usage{
Actual.Values(test.data, target, bn.model)
}
\arguments{
\item{test.data}{a data frame containing all variables. Columns represent variables. Rows represent observations.}

\item{target}{target variable(s)}

\item{bn.model}{Bayesian network model including all variables.}
}
\value{
a list containing actual values of target variable. List has two index. First index represents number of variables.
        Second one represents number of discrete states of variables. For example, if there is one variable that has two different
        discrete states, resulting list would be \code{example.list[1][1]} for first state and \code{example.list[1][2]} for second
        state respectively.
}
\description{
Obtain actual values of target variables.Target variable(s) has different discrete states.
             In each step, a state becomes boolean value `TRUE`, while the rest of the states remain boolean value `FALSE`.
             They are assigned to indexes of list for use in the model validation.
}
\seealso{
\code{\link{ROC_Calculation}}
}
