% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{cond.ent}
\alias{cond.ent}
\title{Conditional Entropy of a Variable}
\usage{
cond.ent(target, evidence.e, bn.model)
}
\arguments{
\item{target}{variable whose conditional entropy to be calculated. Variable type is a character.}

\item{evidence.e}{evidence variable. Variable type is a character.}

\item{bn.model}{Bayesian network model including all variables}
}
\value{
Conditional entropy of target variable
}
\description{
Calculates conditional entropy of a target variable in Bayesian network given other variable(s).
}
\references{
C. E. Shannon, “A Mathematical Theory of Communication,” Bell Syst. Tech. J., vol. 27, no. 3, pp. 379–423, Jul. 1948.
}
\seealso{
\code{\link[gRain]{querygrain}} \code{\link[gRain]{setEvidence}}
}
