% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{correction}
\alias{correction}
\title{Correction Factor}
\usage{
correction(x, correction_factor = 1e-05)
}
\arguments{
\item{x}{a numerical vector}

\item{correction_factor}{correction factor that will be added to value of `0`. The default value is `0.00001`}
}
\value{
a numerical vector
}
\description{
If a numerical vector has a value of `0`, a correction factor is added to that value.
`0` `1` normalization is applied to the vector.
}
\examples{
vector.1 <- c(1,22,4,0)
correction(vector.1)
}
