% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionBayesian.R
\name{Discretization}
\alias{Discretization}
\title{Convert Continuous Variable into Factor}
\usage{
Discretization(vector.d, cutoff.d, states.d = NULL)
}
\arguments{
\item{vector.d}{a vector containing values of continuous variable.}

\item{cutoff.d}{cutting points required to divide the vector into intervals}

\item{states.d}{a vector containing characters which are used to label intervals. Default value is `NULL`. If value is `NULL`,
intervals are labeled as \code{State_1} \code{State_2}...}
}
\value{
a vector containing state labels representing different intervals.
}
\description{
Variables which have continuous values converted into categorical variables which have different intervals.
}
\details{
minimum and maximum values of \code{vector.d} equal to the lowest and highest point of result vector respectively.
         Interval ranging from minimum point in the \code{vector.d} to the leftmost value defined in the \code{cutoff.d} represents first interval.
         Interval ranging from the rightmost value defined in the \code{cutoff.d} to the maximum point in \code{vector.d} represents last interval.
}
\seealso{
\code{\link[bnlearn]{discretize}} \code{\link{cut}}
}
