% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{information.algorithm}
\alias{information.algorithm}
\title{Algorithm for Posterior Distribution of variable}
\usage{
information.algorithm(
  x,
  target,
  query,
  evidence = NULL,
  bn.model,
  num.iter = 1
)
}
\arguments{
\item{x}{a observation vector containing all variables in Bayesian network as characters. Each row in data frame defined as observation vector.}

\item{target}{a character vector containing variable(s) whose posterior distribution to be calculated.}

\item{query}{a character vector containing input variable(s) to be used when calculating the posterior distribution of target variables.}

\item{evidence}{a vector containing observed variables in the Bayesian network. Default value is `NULL`.}

\item{bn.model}{Bayesian network model including all variables.}

\item{num.iter}{an integer that shows number of query variables which will be used during calculation of
posterior distribution of target variable. Default value is `1`.}
}
\value{
a list containing posterior distribution of target variable.
}
\description{
The posterior distribution of the target variable(s) is calculated using the entropy and mutual information values.
}
\seealso{
\code{\link{mut.inf}} \code{\link[gRain]{setEvidence}}
}
