% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_Learning.R
\name{Model_Learning}
\alias{Model_Learning}
\title{Learn Bayesian Network models}
\usage{
Model_Learning(Train_Data, Algorithm.Names = NULL, Structure.List = NULL)
}
\arguments{
\item{Train_Data}{a data frame containing discrete variables and factors used for learning Bayesian model}

\item{Algorithm.Names}{a list of the algorithms used in Bayesian network model learning}

\item{Structure.List}{If there is pre-defined structure of Bayesian network, parameters of it are learned}
}
\value{
List of models
}
\description{
Learn the structure and parameters of Bayesian network either using algorithms
in R or using user-defined algorithms. If the structure of Bayesian network defined in advance,
parameters of it learned and model added to list.
}
\details{
Title
}
