% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{mut.inf}
\alias{mut.inf}
\title{Mutual Information of a Variable}
\usage{
mut.inf(target, evidence, bn.model)
}
\arguments{
\item{target}{variable(s) whose mutual information to be calculated. Variable type is a character.}

\item{evidence}{evidence variable(s). Variable type is a character.}

\item{bn.model}{Bayesian network model including all variables}
}
\value{
a list containing probability distribution of target variable(s), mutual information between target(s) and evidence variable(s),
        and variable(s) which has maximum mutual information
}
\description{
Calculate mutual information value of variable(s) in Bayesian network given other variables
}
\details{
Multiple target and evidence variables can be defined as vectors provided that the variable type is a character. Mutual information between
         target and evidence variables are calculated and returned.
}
\references{
P. Baudot, M. Tapia, D. Bennequin, and J. M. Goaillard, “Topological information data analysis,” Entropy, vol. 21, no. 9, pp. 1–31, 2019.
}
\seealso{
\code{\link[gRain]{querygrain}} \code{\link{entropy}} \cr \code{\link{cond.ent}}
}
