% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionBayesian.R
\name{Predictions}
\alias{Predictions}
\title{Calculate Posterior Distribution of Variable and Store It in List}
\usage{
Predictions(test.data, target, query, bn.model, num.iter = 1)
}
\arguments{
\item{test.data}{a data frame containing all variables. Columns represent variables. Rows represent observations.}

\item{target}{variable(s) whose posterior distribution to be calculated.}

\item{query}{input variable(s) to be used when calculating the posterior distribution of target variables.}

\item{bn.model}{Bayesian network model including all variables.}

\item{num.iter}{an integer that shows number of query variables which will be used during calculation of
posterior distribution of target variable. Default value is `1`.}
}
\value{
a list containing posterior distribution of target variable. List has two index. Posterior distributions
        obtained for each target variable and each Bayesian network model. First index is for number of variables
        second one is for number of Bayesian network models.
}
\description{
posterior distribution of target variable(s) is calculated. \code{\link{information.algorithm}} is used
             in order to obtain distributions. Obtained distributions are stored in lists.
}
\seealso{
\code{\link{mut.inf}} \code{\link{information.algorithm}}
}
