% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionBayesian.R
\name{ROC_Calculation}
\alias{ROC_Calculation}
\title{Obtain ROC curve and AUC}
\usage{
ROC_Calculation(predictions, a.Values)
}
\arguments{
\item{predictions}{posterior Distributions of target variable(s).}

\item{a.Values}{actual values of target variable(s).}
}
\value{
a list containing ROC curve and AUC. List has two index. First index contains AUC values of model(s). Second index contains
        contains ROC curve(s) of model(s).
}
\description{
This function created in order to obtain ROC curve and AUC values of model(s). Posterior Distributions of variable(s)
             is compared to actual values of variable(s).
}
\details{
Posterior distributions of target variable(s) obtained through \code{\link{Predictions}} function is compared to
         actual values of target variable(s) obtained through \code{\link{Actual.Values}} function.
}
\seealso{
\code{\link[pROC]{roc}} \code{\link[pROC]{auc}}
}
