% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionBayesian.R
\name{Model_Learning}
\alias{Model_Learning}
\title{Learn Bayesian Network models}
\usage{
Model_Learning(train.data, algorithm.names = NULL, structure.list = NULL)
}
\arguments{
\item{train.data}{a data frame containing discrete variables and factors used for learning Bayesian model}

\item{algorithm.names}{a vector containing the algorithms used in Bayesian network model learning. The default value is `NULL`}

\item{structure.list}{If there is pre-defined structure of Bayesian network, parameters of it are learned. The default value is `NULL`}
}
\value{
List of Bayesian network models. Each element of the list is represented by the name of the algorithm.
If the list element is the predefined model, it is named \code{Predefined_Model_1}, \code{Predefined_Model_2}...
}
\description{
Learn the structure and parameters of Bayesian network either using algorithms
in R or using user-defined algorithms. If the structure of Bayesian network defined in advance,
parameters of it learned and model added to list.
}
\details{
Available algorithms in R could be used to learn structure and parameters. See \code{\link[bnlearn]{structure-learning}}
for available algorithms. Also user-defined algorithms could be defined and used. If the structure(s) of Bayesian network defined in advance
parameters of it learned and model(s) added to list.
}
\seealso{
\code{\link[bnlearn]{bn.fit}} \code{\link[bnlearn]{as.grain}}
}
